storage: move testing functions to their own file
Also removes unused argument.
